Replace remaining mentions of ASCII with UTF-8
The previous commit fixed the actual bug, and all these remaining
mentions of ASCII are probably guaranteed to be ASCII (not sure about
object type tags, but a scenario where a non-ASCII one is encountered is
unrealistic).

But it’s nicer to deal with UTF-8 everywhere (which is, after all,
Python 3’s default encoding) and never mention ASCII anywhere it’s not
expressly needed, so here’s what I’m going to call linting.