Update documentation and to-do list help
Deleted lines are no longer treated as index commands. Mention that
the number of lines must remain identical and that index commands must
come last. Adapt 'fixup' and 'squash' to mirror 'git rebase', so that
the 'fixup' help in the todo list mentions that the message is dropped.
Misc rephrasing.