Query editor and add syntax hint string to edit buffer
Some editors come with syntax highlighting for rebase and commit message
buffers and with a way to read hints suggesting a specific highlighting.
Add such a hint before running the editor. The format for 'git commit'
is completely identical and the 'git-rebase-todo' one is quite close to
and also work with the 'git-revise-todo' format.

("Some" is currently a lie, only works with vim.)