Fix order of fixup commits with autosquash
When adding a new fixup to commit X, we want to add it after all previous
fixup commits that target X, and after any fixups to those commits, and so on.

The cycle detection might be a bit over-the-top, since the logic only allows
to create self cycles with a message starting with "fixup! fixup!".
If we only allowed to fixup previous commits in the todo list, then there would
not be cycles by design, that could simplify this patch.

Fixes #69