Support git-rerere to record and replay conflict resolution (#75)
This adds basic support for [git-rerere].
Whenever a merge conflict in a file is successfully resolved by the user,
the conflict (the preimage) and the merged file (the postimage) are recorded.
When resolving the same conflict again, the recorded resolution is replayed,
unless it doesn't merge cleanly with the current preimage.

The resolutions are shared with the ones that git-rerere records;
both save them in .git/rr-cache/<conflict-id>.

My motivation is that I'm using git-revise in a script that sometimes needs
to resolve the same conflicts over and over again. By caching conflict
resolutions I only need to do that once.

This feature is disabled by default.  Use it by enabling revise.rerere (or
rerere.enabled) _and_ rerere.autoUpdate. In future, we plan to also support
use without rerere.autoUpdate.

Upstream git-rerere can remember multiple resolutions for each conflict,
so-called "variants".  This is not yet supported by this patch - we only
keep the last resolution for each conflict around. This should probably be
added in a future patch.

[git-rerere]: <https://git-scm.com/docs/git-rerere/>