Disable Windows CI and run CI on PRs (#89)
* Disable Windows CI for now

I tried to use Cygwin but no luck yet.  Running bash would always
result into this error:

	Windows Subsystem for Linux has no installed distributions.

Which can probably be fixed easily, but it's not urgent. Just disable
Windows for now to avoid some noise.

* Run CI on pull requests

This seems to be the standard - it makes it easy to see if a PR
breaks tests.