lint: use `isort` to normalize import order
Note that `pylint` already depends on `isort`. Using it makes import
sorting unambiguous and automatic.

It's a little ambiguous whether it should be in the "lint" or "format"
tox command, or another one altogether. It's included in "lint" here,
because pylint already uses isort internally to catch coarse-grained
"wrong-import-order" warnings.