test.edit: Async editor (#113)
This fixes some race conditions in the editing server by introducing a
blocking queue of edit requests. This means that if the launched
git-revise invokes the editor before the test invokes `next_file()`, the
request will be pending.

This also splits up "EditorFile" (the mutable/inspectable edit data that
the test can manipulate) from the HTTP request itself. Beyond separation
of those concerns, a side effect is that this fixes pylint's ability to
disentangle the expected type of the `next_file` context manager.
Previously the mutual modification between `Editor` aned `EditFile`
forced it to give up.