edit: Run editor from cwd, not parent of file (#118)
* edit: Run editor from cwd, not parent of file

This is how git-rebase does it, and doing so makes it easier to refer to
local files/notes when composing edit messages, rather than being
plopped into a random tmpdir.

* edit+merge: show relative paths, not anchored at /

This is more consistent with how git-rebase does it. Anchoring at `/` is
misleading because it displays an absolute path that usually does not
exist!