Support pathspecs in --cut
Sometimes I want to extract a small number of changes from a commit
that touches loads of files. Since the cut mode prompts about every
single file, there is no convenient way to do this.

Teach --cut about pathspec arguments.  This allows to use

	git revise --cut my-commit -- path/to/my/file
	a

to extract to a separate commit all changes in the given file.

In future we could support pathspecs in other subcommands.